Performance improvements and bug fixes
The game loop was consolidated to a parent container, which allowed me to remove 1 game loop and improves performance. 
In the previous version the enemy would get "stuck" on the wall due to how movement was calculated. There is now a time check to ensure that the enemy has time to move far enough away without triggering a direction change. 
The refactor needs some work still including collision detection and the explosion animation when the enemy hits the player.